Refactor managers
Separate Trigger classes so that they work with generics
