+++ bukkit/src/test/java/js/executor/AbstractTestExecutors.java
@@ -5,7 +5,7 @@ package js.executor;
 import io.github.wysohn.triggerreactor.core.bridge.IInventory;
 import io.github.wysohn.triggerreactor.core.bridge.entity.IPlayer;
 import io.github.wysohn.triggerreactor.core.main.TriggerReactorCore;
-import io.github.wysohn.triggerreactor.core.manager.trigger.AbstractInventoryTriggerManager;
+import io.github.wysohn.triggerreactor.core.manager.trigger.inventory.AbstractInventoryTriggerManager;
 import js.AbstractTestJavaScripts;
 import js.ExecutorTest;
 import js.JsTest;
