+++ core/src/main/java/io/github/wysohn/triggerreactor/core/manager/trigger/area/AbstractAreaTriggerManager.java
@@ -14,17 +14,15 @@
  *     You should have received a copy of the GNU General Public License
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *******************************************************************************/
-package io.github.wysohn.triggerreactor.core.manager.trigger;
+package io.github.wysohn.triggerreactor.core.manager.trigger.area;
 
 import io.github.wysohn.triggerreactor.core.bridge.entity.IEntity;
 import io.github.wysohn.triggerreactor.core.main.TriggerReactorCore;
 import io.github.wysohn.triggerreactor.core.manager.location.Area;
 import io.github.wysohn.triggerreactor.core.manager.location.SimpleChunkLocation;
 import io.github.wysohn.triggerreactor.core.manager.location.SimpleLocation;
-import io.github.wysohn.triggerreactor.core.script.interpreter.Interpreter;
+import io.github.wysohn.triggerreactor.core.manager.trigger.AbstractTaggedTriggerManager;
 import io.github.wysohn.triggerreactor.tools.FileUtil;
-import io.github.wysohn.triggerreactor.tools.StringUtils;
-import io.github.wysohn.triggerreactor.core.manager.trigger.AbstractTriggerManager.Trigger;
 
 import java.io.File;
 import java.io.FileFilter;
@@ -36,7 +34,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 
-public abstract class AbstractAreaTriggerManager extends AbstractTaggedTriggerManager<AbstractAreaTriggerManager.AreaTrigger> {
+public abstract class AbstractAreaTriggerManager extends AbstractTaggedTriggerManager<AreaTrigger> {
     protected static final String SMALLEST = "Smallest";
     protected static final String LARGEST = "Largest";
     protected static final String SYNC = "Sync";
@@ -414,223 +412,4 @@ public abstract class AbstractAreaTriggerManager extends AbstractTaggedTriggerMa
         FileUtil.delete(new File(folder, trigger.getTriggerName()));
     }
 
-    public static class AreaTrigger extends Trigger {
-        final Area area;
-        final File folder;
-
-        private EnterTrigger enterTrigger;
-        private ExitTrigger exitTrigger;
-
-        private Map<UUID, WeakReference<IEntity>> trackedEntities = new ConcurrentHashMap<>();
-
-        public AreaTrigger(Area area, File folder, String name) {
-            super(name, null, null);
-            this.area = area;
-            this.folder = folder;
-        }
-
-        public Area getArea() {
-            return area;
-        }
-
-        //we don't need interpreter for area trigger but enter and exit trigger
-        @Override
-        protected Interpreter initInterpreter(Map<String, Object> scriptVars) {
-            return null;
-        }
-
-        private EventType type = null;
-
-        public void activate(Object e, Map<String, Object> scriptVars, EventType type) {
-            this.type = type;
-
-            super.activate(e, scriptVars);
-        }
-
-        //intercept and pass interpretation to appropriate trigger
-        @Override
-        protected void startInterpretation(Object e, Map<String, Object> scriptVars, Interpreter interpreter, boolean sync) {
-            switch (type) {
-                case ENTER:
-                    if (getEnterTrigger() != null)
-                        getEnterTrigger().activate(e, scriptVars);
-                    break;
-                case EXIT:
-                    if (getExitTrigger() != null)
-                        getExitTrigger().activate(e, scriptVars);
-                    break;
-                default:
-                    throw new RuntimeException("Unknown area event type " + type);
-            }
-        }
-
-        @Override
-        public Trigger clone() {
-            return null;
-        }
-
-        @Override
-        protected String getTimingId() {
-            return StringUtils.dottedPath(super.getTimingId(), area.toString());
-        }
-
-        @Override
-        public String toString() {
-            return super.toString() + "{" +
-                    "area=" + area +
-                    '}';
-        }
-
-        void setEnterTrigger(String script, File triggerFile) throws TriggerInitFailedException {
-            enterTrigger = new EnterTrigger(this, triggerFile, script);
-        }
-
-        public void setEnterTrigger(String script) throws TriggerInitFailedException {
-            File triggerFile = getTriggerFile(folder, "Enter", true);
-            setEnterTrigger(script, triggerFile);
-        }
-
-        void setExitTrigger(String script, File triggerFile) throws TriggerInitFailedException {
-            exitTrigger = new ExitTrigger(this, triggerFile, script);
-        }
-
-        public void setExitTrigger(String script) throws TriggerInitFailedException {
-            File triggerFile = getTriggerFile(folder, "Enter", true);
-            setExitTrigger(script, triggerFile);
-        }
-
-        public EnterTrigger getEnterTrigger() {
-            return enterTrigger;
-        }
-
-        public void setEnterTrigger(EnterTrigger enterTrigger) {
-            this.enterTrigger = enterTrigger;
-        }
-
-        public ExitTrigger getExitTrigger() {
-            return exitTrigger;
-        }
-
-        public void setExitTrigger(ExitTrigger exitTrigger) {
-            this.exitTrigger = exitTrigger;
-        }
-
-        public void addEntity(IEntity entity) {
-            WeakReference<IEntity> ref = new WeakReference<>(entity);
-            this.trackedEntities.put(entity.getUniqueId(), ref);
-        }
-
-        public void removeEntity(UUID uuid) {
-            this.trackedEntities.remove(uuid);
-        }
-
-        public IEntity getEntity(UUID uuid) {
-            WeakReference<IEntity> ref = this.trackedEntities.get(uuid);
-            if (ref == null)
-                return null;
-
-            IEntity entity = ref.get();
-            //just remove it as it's got garbage-collected.
-            if (entity == null) {
-                this.trackedEntities.remove(uuid);
-            }
-
-            return entity;
-        }
-
-        public List<IEntity> getEntities() {
-            List<IEntity> entities = new ArrayList<>();
-
-            Set<UUID> remove = new HashSet<>();
-            for (Entry<UUID, WeakReference<IEntity>> entry : this.trackedEntities.entrySet()) {
-                WeakReference<IEntity> ref = entry.getValue();
-                IEntity entity = ref.get();
-                if (entity != null) {
-                    entities.add(entity);
-                } else {
-                    remove.remove(entry.getKey());
-                }
-            }
-
-            for (UUID uuid : remove) {
-                this.trackedEntities.remove(uuid);
-            }
-
-            return entities;
-        }
-
-        public static class EnterTrigger extends Trigger {
-            private final AreaTrigger areaTrigger;
-
-            public EnterTrigger(AreaTrigger areaTrigger, File file, String script) throws TriggerInitFailedException {
-                super(areaTrigger.triggerName, file, script);
-                this.areaTrigger = areaTrigger;
-
-                init();
-            }
-
-            @Override
-            protected String getTimingId() {
-                return StringUtils.dottedPath(areaTrigger.getTimingId(), "Enter");
-            }
-
-            @Override
-            public boolean isSync() {
-                return areaTrigger.isSync();
-            }
-
-            @Override
-            public void setSync(boolean sync) {
-                areaTrigger.setSync(sync);
-            }
-
-            @Override
-            public Trigger clone() {
-                try {
-                    return new EnterTrigger(areaTrigger, file, script);
-                } catch (TriggerInitFailedException e) {
-                    e.printStackTrace();
-                }
-                return null;
-            }
-
-        }
-
-        public static class ExitTrigger extends Trigger {
-            private final AreaTrigger areaTrigger;
-
-            public ExitTrigger(AreaTrigger areaTrigger, File file, String script) throws TriggerInitFailedException {
-                super(areaTrigger.getTriggerName(), file, script);
-                this.areaTrigger = areaTrigger;
-
-                init();
-            }
-
-            @Override
-            protected String getTimingId() {
-                return StringUtils.dottedPath(areaTrigger.getTimingId(), "Exit");
-            }
-
-            @Override
-            public boolean isSync() {
-                return areaTrigger.isSync();
-            }
-
-            @Override
-            public void setSync(boolean sync) {
-                areaTrigger.setSync(sync);
-            }
-
-            @Override
-            public Trigger clone() {
-                try {
-                    return new ExitTrigger(areaTrigger, file, script);
-                } catch (TriggerInitFailedException e) {
-                    e.printStackTrace();
-                }
-                return null;
-            }
-
-        }
-    }
 }
