+++ core/src/main/java/io/github/wysohn/triggerreactor/core/manager/trigger/command/CommandTrigger.java
@@ -1,0 +1,59 @@
+package io.github.wysohn.triggerreactor.core.manager.trigger.command;
+
+import io.github.wysohn.triggerreactor.core.manager.trigger.AbstractTriggerManager;
+
+import java.io.File;
+import java.util.Arrays;
+
+public class CommandTrigger extends AbstractTriggerManager.Trigger {
+    String[] permissions = new String[0];
+    String[] aliases = new String[0];
+
+    public CommandTrigger(String name, File file, String script) throws AbstractTriggerManager.TriggerInitFailedException {
+        super(name, file, script);
+
+        init();
+    }
+
+    public String[] getPermissions() {
+        return permissions;
+    }
+
+    public void setPermissions(String[] permissions) {
+        if (permissions == null) {
+            this.permissions = new String[0];
+        } else {
+            this.permissions = permissions;
+        }
+    }
+
+    public String[] getAliases() {
+        return aliases;
+    }
+
+    public void setAliases(String[] aliases) {
+        if (aliases == null) {
+            this.aliases = new String[0];
+        } else {
+            this.aliases = aliases;
+        }
+    }
+
+    @Override
+    public AbstractTriggerManager.Trigger clone() {
+        try {
+            return new CommandTrigger(triggerName, file, getScript());
+        } catch (AbstractTriggerManager.TriggerInitFailedException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    @Override
+    public String toString() {
+        return super.toString() + "{" +
+                "permissions=" + Arrays.toString(permissions) +
+                ", aliases=" + Arrays.toString(aliases) +
+                '}';
+    }
+}
