+++ core/src/main/java/io/github/wysohn/triggerreactor/core/manager/trigger/inventory/AbstractInventoryTriggerManager.java
@@ -14,35 +14,31 @@
  *     You should have received a copy of the GNU General Public License
  *     along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *******************************************************************************/
-package io.github.wysohn.triggerreactor.core.manager.trigger;
+package io.github.wysohn.triggerreactor.core.manager.trigger.inventory;
 
 import io.github.wysohn.triggerreactor.core.bridge.IInventory;
 import io.github.wysohn.triggerreactor.core.bridge.IItemStack;
 import io.github.wysohn.triggerreactor.core.bridge.entity.IPlayer;
 import io.github.wysohn.triggerreactor.core.main.TriggerReactorCore;
-import io.github.wysohn.triggerreactor.core.script.interpreter.Interpreter;
+import io.github.wysohn.triggerreactor.core.manager.trigger.AbstractTriggerManager;
 import io.github.wysohn.triggerreactor.core.script.lexer.LexerException;
 import io.github.wysohn.triggerreactor.core.script.parser.ParserException;
 import io.github.wysohn.triggerreactor.tools.FileUtil;
-import io.github.wysohn.triggerreactor.tools.timings.Timings;
-import io.github.wysohn.triggerreactor.core.manager.trigger.AbstractTriggerManager.Trigger;
 
 import java.io.File;
 import java.io.FileFilter;
 import java.io.IOException;
-import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.concurrent.ConcurrentHashMap;
 
-public abstract class AbstractInventoryTriggerManager extends AbstractTriggerManager<AbstractInventoryTriggerManager.InventoryTrigger> {
+public abstract class AbstractInventoryTriggerManager extends AbstractTriggerManager<InventoryTrigger> {
     protected static final String ITEMS = "Items";
     protected static final String SIZE = "Size";
 
-    private final static Map<IInventory, InventoryTrigger> inventoryMap = new ConcurrentHashMap<>();
-
-    private final Map<IInventory, Map<String, Object>> inventorySharedVars = new ConcurrentHashMap<>();
+    final static Map<IInventory, InventoryTrigger> inventoryMap = new ConcurrentHashMap<>();
+    final Map<IInventory, Map<String, Object>> inventorySharedVars = new ConcurrentHashMap<>();
 
     @Override
     public void reload() {
@@ -233,7 +229,7 @@ public abstract class AbstractInventoryTriggerManager extends AbstractTriggerMan
 
     @Override
     protected void deleteInfo(InventoryTrigger trigger) {
-        FileUtil.delete(new File(trigger.file.getParent(), trigger.getTriggerName() + ".yml"));
+        FileUtil.delete(new File(trigger.getFile().getParent(), trigger.getTriggerName() + ".yml"));
         super.deleteInfo(trigger);
     }
 
@@ -276,63 +272,6 @@ public abstract class AbstractInventoryTriggerManager extends AbstractTriggerMan
 
     public AbstractInventoryTriggerManager(TriggerReactorCore plugin, File tirggerFolder) {
         super(plugin, tirggerFolder);
-    }
-
-    public static class InventoryTrigger extends Trigger {
-        public static final int MAXSIZE = 6 * 9;
-
-        final IItemStack[] items;
-
-        private InventoryTrigger(String name, String script, File file, IItemStack[] items) throws TriggerInitFailedException {
-            super(name, file, script);
-            this.items = items;
-
-            init();
-        }
-
-        public InventoryTrigger(int size, String name, Map<Integer, IItemStack> items, File file, String script) throws TriggerInitFailedException {
-            super(name, file, script);
-            if (size < 9 || size % 9 != 0)
-                throw new IllegalArgumentException("Inventory Trigger size should be multiple of 9!");
-
-            if (size > MAXSIZE)
-                throw new IllegalArgumentException("Inventory Size cannot be larger than " + MAXSIZE);
-
-            this.items = new IItemStack[size];
-
-            for (Map.Entry<Integer, IItemStack> entry : items.entrySet()) {
-                this.items[entry.getKey()] = entry.getValue();
-            }
-
-            init();
-        }
-
-        @Override
-        protected void start(Timings.Timing timing, Object e, Map<String, Object> scriptVars, Interpreter interpreter,
-                             boolean sync) {
-            try {
-                interpreter.startWithContextAndInterrupter(e,
-                        TriggerReactorCore.getInstance().createInterrupterForInv(e, interpreter, cooldowns, inventoryMap),
-                        timing);
-            } catch (Exception ex) {
-                TriggerReactorCore.getInstance().handleException(e,
-                        new Exception("Error occurred while processing Trigger [" + getTriggerName() + "]!", ex));
-            }
-        }
-
-        @Override
-        public Trigger clone() {
-            try {
-                return new InventoryTrigger(triggerName, script, file, items);
-            } catch (TriggerInitFailedException e) {
-                e.printStackTrace();
-            }
-            return null;
-        }
-
-        public IItemStack[] getItems() {
-            return items;
-        }
     }
 
 }
